Update local smoke tests script to support multiple OpenAPI spec files

+++ test/smoke-tests.ps1
@@ -23,37 +23,55 @@ function RunTests {
         $Parallel = $false
     )
 
+    $filenames = @(
+        # "api-with-examples",
+        # "callback-example",
+        # "link-example",
+        # "uber",
+        # "uspto",
+        # "petstore-expanded",
+        # "petstore-minimal",
+        # "petstore-simple",
+        # "petstore-with-external-docs",
+        "petstore"
+    )
     "v2.0", "v3.0" | ForEach-Object {
         $version = $_
-        "json", "yaml" | ForEach-Object {
+        "json", "yaml" | ForEach-Object {            
             $format = $_
-            Remove-Item ./**/*Output.cs -Force
-            Copy-Item ./OpenAPI/$version/petstore.$format ./openapi.$format
+            $filenames | ForEach-Object {
+                $filename = "./OpenAPI/$version/$_.$format"
+                $exists = Test-Path -Path $filename -PathType Leaf
+                if ($exists -eq $true) {
+                    Remove-Item ./**/*Output.cs -Force
+                    Copy-Item $filename ./openapi.$format
 
-            dotnet run `
-                --project ../src/Refitter/Refitter.csproj `
-                ./openapi.$format `
-                --namespace GeneratedCode
+                    dotnet run `
+                        --project ../src/Refitter/Refitter.csproj `
+                        ./openapi.$format `
+                        --namespace GeneratedCode
 
-            Copy-Item "./Output.cs" "./GeneratedCode/Net7/Output.cs" -Force
-            Copy-Item "./Output.cs" "./GeneratedCode/Net7/Output.cs" -Force
-            Copy-Item "./Output.cs" "./GeneratedCode/Net6/Output.cs" -Force
-            Copy-Item "./Output.cs" "./GeneratedCode/Net48/Output.cs" -Force
-            Copy-Item "./Output.cs" "./GeneratedCode/Net481/Output.cs" -Force
-            Copy-Item "./Output.cs" "./GeneratedCode/Net472/Output.cs" -Force
-            Copy-Item "./Output.cs" "./GeneratedCode/Net462/Output.cs" -Force
-            Copy-Item "./Output.cs" "./GeneratedCode/NetStandard20/Output.cs" -Force
-            Copy-Item "./Output.cs" "./GeneratedCode/NetStandard21/Output.cs" -Force
-            Remove-Item "./Output.cs" -Force
+                    Copy-Item "./Output.cs" "./GeneratedCode/Net7/Output.cs" -Force
+                    Copy-Item "./Output.cs" "./GeneratedCode/Net7/Output.cs" -Force
+                    Copy-Item "./Output.cs" "./GeneratedCode/Net6/Output.cs" -Force
+                    Copy-Item "./Output.cs" "./GeneratedCode/Net48/Output.cs" -Force
+                    Copy-Item "./Output.cs" "./GeneratedCode/Net481/Output.cs" -Force
+                    Copy-Item "./Output.cs" "./GeneratedCode/Net472/Output.cs" -Force
+                    Copy-Item "./Output.cs" "./GeneratedCode/Net462/Output.cs" -Force
+                    Copy-Item "./Output.cs" "./GeneratedCode/NetStandard20/Output.cs" -Force
+                    Copy-Item "./Output.cs" "./GeneratedCode/NetStandard21/Output.cs" -Force
+                    Remove-Item "./Output.cs" -Force
 
-            Write-Host "`r`nBuilding $_`r`n"
-            $process = Start-Process "dotnet" `
-                -Args "build ./GeneratedCode/Generated.sln" `
-                -NoNewWindow `
-                -PassThru
-            $process | Wait-Process
-            if ($process.ExitCode -ne 0) {
-                throw "Build Failed!"
+                    Write-Host "`r`nBuilding $_`r`n"
+                    $process = Start-Process "dotnet" `
+                        -Args "build ./GeneratedCode/Generated.sln" `
+                        -NoNewWindow `
+                        -PassThru
+                    $process | Wait-Process
+                    if ($process.ExitCode -ne 0) {
+                        throw "Build Failed!"
+                    }                
+                }
             }
         }
     }
